CLOUD-524: Refresh integration test suit configurations
