+++ integrationtest/src/main/java/com/sequenceiq/it/cloudbreak/config/CloudbreakITConfig.java
@@ -3,12 +3,11 @@ package com.sequenceiq.it.cloudbreak.config;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-import com.sequenceiq.it.cloudbreak.TemplateAdditionParser;
+import com.sequenceiq.it.cloudbreak.TemplateAdditionHelper;
 
 @Configuration
 public class CloudbreakITConfig {
-    @Bean
-    TemplateAdditionParser templateAdditionParser() {
-        return new TemplateAdditionParser();
+    @Bean TemplateAdditionHelper templateAdditionHelper() {
+        return new TemplateAdditionHelper();
     }
 }
