+++ integrationtest/src/main/resources/testng-aws-cluster.xml
@@ -1,6 +1,6 @@
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
 
-<suite name="Azure Test with precreated resources">
+<suite name="Aws Test with precreated resources">
     <!--
     <parameter name="uaaServer" value="COULD BE SPECIFIED" />
     <parameter name="uaaUser" value="COULD BE SPECIFIED" />
@@ -8,38 +8,20 @@
     <parameter name="cloudbreakServer" value="COULD BE SPECIFIED" />
     -->
     <!-- resources -->
-    <parameter name="credentialName" value="itazurecredential" />
+    <parameter name="credentialName" value="it-aws-credential" />
     <parameter name="blueprintName" value="multi-node-hdfs-yarn" />
+    <parameter name="instanceGroups" value="it-aws-master,master,1;it-aws-slave,slave_1,3" />
+    <parameter name="hostGroups" value="master,master;slave_1,slave_1" />
     <test name="init">
         <classes>
             <class name="com.sequenceiq.it.TestSuiteInitializer" />
             <class name="com.sequenceiq.it.cloudbreak.CloudbreakTestSuiteInitializer" />
         </classes>
     </test>
-    <test name="create master template">
-        <parameter name="azureTemplateName" value="it-azure-master-created" />
-        <parameter name="azureVmType" value="A5" />
-        <parameter name="azureVolumeCount" value="1" />
-        <parameter name="azureVolumeSize" value="30" />
-        <parameter name="templateAdditions" value="master,1" />
-        <classes>
-            <class name="com.sequenceiq.it.cloudbreak.AzureTemplateCreationTest" />
-        </classes>
-    </test>
-    <test name="create slave template">
-        <parameter name="azureTemplateName" value="it-azure-slave-created" />
-        <parameter name="azureVmType" value="A5" />
-        <parameter name="azureVolumeCount" value="2" />
-        <parameter name="azureVolumeSize" value="10" />
-        <parameter name="templateAdditions" value="slave_1,3" />
-        <classes>
-            <class name="com.sequenceiq.it.cloudbreak.AzureTemplateCreationTest" />
-        </classes>
-    </test>
     <test name="create cluster">
-        <parameter name="stackName" value="it-azure-stack" />
-        <parameter name="region" value="NORTH_EUROPE" />
-        <parameter name="clusterName" value="it-azure-cluster" />
+        <parameter name="stackName" value="it-aws-stack" />
+        <parameter name="region" value="EU_WEST_1" />
+        <parameter name="clusterName" value="it-aws-cluster" />
         <classes>
             <class name="com.sequenceiq.it.cloudbreak.StackCreationTest" />
             <class name="com.sequenceiq.it.cloudbreak.ClusterCreationTest" />
