+++ integrationtest/src/main/resources/testng-aws-full.xml
@@ -7,7 +7,6 @@
     <parameter name="uaaPassword" value="COULD BE SPECIFIED" />
     <parameter name="cloudbreakServer" value="COULD BE SPECIFIED" />
     -->
-    <parameter name="hostGroups" value="master,master;slave_1,slave_1" />
     <test name="init">
         <classes>
             <class name="com.sequenceiq.it.TestSuiteInitializer" />
