+++ integrationtest/src/main/resources/testng-aws-pre.xml
@@ -10,10 +10,6 @@
     <!-- resources -->
     <parameter name="credentialName" value="it-aws-credential" />
     <parameter name="blueprintName" value="multi-node-hdfs-yarn" />
-    <!-- with existing cloud provider templates
-    <parameter name="instanceGroups" value="it-aws-master,master,1;it-aws-slave,slave_1,3" />
-    -->
-    <parameter name="hostGroups" value="master,master;slave_1,slave_1" />
     <test name="init">
         <classes>
             <class name="com.sequenceiq.it.TestSuiteInitializer" />
