+++ integrationtest/src/main/resources/testng-azure-cluster.xml
@@ -10,30 +10,12 @@
     <!-- resources -->
     <parameter name="credentialName" value="itazurecredential" />
     <parameter name="blueprintName" value="multi-node-hdfs-yarn" />
+    <parameter name="instanceGroups" value="it-azure-master,master,1;it-azure-slave,slave_1,3" />
+    <parameter name="hostGroups" value="master,master;slave_1,slave_1" />
     <test name="init">
         <classes>
             <class name="com.sequenceiq.it.TestSuiteInitializer" />
             <class name="com.sequenceiq.it.cloudbreak.CloudbreakTestSuiteInitializer" />
-        </classes>
-    </test>
-    <test name="create master template">
-        <parameter name="azureTemplateName" value="it-azure-master-created" />
-        <parameter name="azureVmType" value="A5" />
-        <parameter name="azureVolumeCount" value="1" />
-        <parameter name="azureVolumeSize" value="30" />
-        <parameter name="templateAdditions" value="master,1" />
-        <classes>
-            <class name="com.sequenceiq.it.cloudbreak.AzureTemplateCreationTest" />
-        </classes>
-    </test>
-    <test name="create slave template">
-        <parameter name="azureTemplateName" value="it-azure-slave-created" />
-        <parameter name="azureVmType" value="A5" />
-        <parameter name="azureVolumeCount" value="2" />
-        <parameter name="azureVolumeSize" value="10" />
-        <parameter name="templateAdditions" value="slave_1,3" />
-        <classes>
-            <class name="com.sequenceiq.it.cloudbreak.AzureTemplateCreationTest" />
         </classes>
     </test>
     <test name="create cluster">
